wip: kafka listener
so far I've gotten to do what I expect however, I will most likely need to build out an email service to manage this properly. This will send the corresponding data to the email service and then the email service will take care of sending it to the user with all of the styles.